requirements.txt
einops==0.8.1
fastai==2.8.2
h5py==3.13.0
huggingface_hub==0.30.1
ipython==8.12.3
lpips==0.1.4
matplotlib==3.10.3
natsort==8.4.0
natten==0.17.5
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
ptflops==0.7.4
Requests==2.32.3
scipy==1.15.3
six==1.17.0
six==1.16.0
skimage==0.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
ultralytics_thop==2.0.14
xformers==0.0.30
yacs==0.1.8
